+++ src/types/Master.ts
@@ -6,19 +6,15 @@ export type MasterConstants = {
     ASSET_PRICE_SCALE: bigint,
     ASSET_RESERVE_FACTOR_SCALE: bigint,
     ASSET_LIQUIDATION_RESERVE_FACTOR_SCALE: bigint,
+    ASSET_LIQUIDATION_THRESHOLD_SCALE: bigint,
+    ASSET_LIQUIDATION_BONUS_SCALE: bigint,
     ASSET_ORIGINATION_FEE_SCALE: bigint
 };
 
-export type PoolAssetConfig = (PoolTonAssetConfig | PoolJettonAssetConfig) & {
-    name: string;
-};
 export type PoolAssetsConfig = PoolAssetConfig[];
 
-export type PoolTonAssetConfig = {
-    assetId: bigint;
-}
-
-export type PoolJettonAssetConfig = {
+export type PoolAssetConfig = {
+    name: string;
     assetId: bigint;
     jettonMasterAddress: Address;
     jettonWalletCode: Cell;
@@ -115,3 +111,8 @@ export type MasterData = {
         borrow: Dictionary<bigint, number>;
     };
 };
+
+export type AgregatedBalances = {
+    totalBorrow: bigint;
+    totalSupply: bigint;
+}